Fixed the panic error.
```bash
panic: runtime error: invalid memory address or nil pointer dereference
[signal SIGSEGV: segmentation violation code=0x1 addr=0x28 pc=0x42db55b]
```

The infix operator was designed to only evaluate string type objects. But the bug was in the conditional. The string type object was evaluated only if the left type object was string. It should be evaluated if the left and right type object was string.