Add regexp support to evaluateExpressions
This pull request adds support for regular expression values to the evaluateExpressions function. It checks if the evaluated expression is a string containing a regular expression pattern, and if it is, it compiles the regular expression and stores it in the result slice. If there is an error compiling the regular expression, it returns an error object. This allows the evaluateExpressions function to process regular expression values in a more straightforward and efficient manner.